% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{AD_DP}
\alias{AD_DP}
\title{Obtain editing levels for results
Reformats results from Sprint file}
\usage{
AD_DP(RES.table)
}
\arguments{
\item{RES.table}{data frame of RES table from Sprint obtained from read.delim() function}
}
\description{
Obtain editing levels for results
Reformats results from Sprint file
}
