% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{annotate_regions}
\alias{annotate_regions}
\title{A function to intersect user region data with annotation data
Taken from annotatr package
Annotate genomic regions to selected genomic annotations while preserving the data associated with the genomic regions.}
\usage{
annotate_regions(
  regions,
  annotations = hg38_annotations,
  minoverlap = 1L,
  ignore.strand = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{regions}{The GRanges object to annotate}

\item{annotations}{The annotations to overlap with}

\item{minoverlap}{A scalar, positive integer, indicating the minimum required overlap of regions with annotations.}

\item{ignore.strand}{Logical indicating whether strandedness should be respected in findOverlaps(). Default FALSE.}

\item{quiet}{Print progress messages (FALSE) or not (TRUE).

' @export}
}
\description{
A function to intersect user region data with annotation data
Taken from annotatr package
Annotate genomic regions to selected genomic annotations while preserving the data associated with the genomic regions.
}
