% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{filter_annotated_results}
\alias{filter_annotated_results}
\title{A function for filtering annotated Sprint results by editing ratio and/or number of supporting reads}
\usage{
filter_annotated_results(input.sprint, editing_ratio = 0, supporting_reads = 1)
}
\arguments{
\item{input.sprint}{The Sprint data frame to input}

\item{editing_ratio}{The minimum editing ratio to have}

\item{supporting_reads}{The minimum number of supporting reads needed}
}
\description{
A function for filtering annotated Sprint results by editing ratio and/or number of supporting reads
}
