% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{filter_by_transitions}
\alias{filter_by_transitions}
\title{A function for filtering annotated Sprint results by type of transitions}
\usage{
filter_by_transitions(input.sprint, transitions_list = c("AG", "TC"))
}
\arguments{
\item{input.sprint}{The Sprint data frame to input}

\item{transitions_list}{The list of transitions you want}
}
\description{
A function for filtering annotated Sprint results by type of transitions
}
