% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{getGeneMatrix}
\alias{getGeneMatrix}
\title{A function for obtaining a gene matrix from a list of tables of identified genes}
\usage{
getGeneMatrix(list_table_identified_genes)
}
\arguments{
\item{list_table_identified_genes}{A list of tables from table_identified_genes() with names of tables being sample names}
}
\description{
A function for obtaining a gene matrix from a list of tables of identified genes
}
