% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{partition_files}
\alias{partition_files}
\title{A function for annotating the Sprint results with genomic regions and repeats (allows for partition)}
\usage{
partition_files(
  input.file.path,
  partition.path = "partitions",
  partition = 10000
)
}
\arguments{
\item{input.file.path}{The input file path}

\item{partition.path}{Directory to export partitions to}

\item{partition}{the number of RES to partition by}
}
\description{
A function for annotating the Sprint results with genomic regions and repeats (allows for partition)
}
