% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{RES.name}
\alias{RES.name}
\title{Obtain site names for RES results}
\usage{
RES.name(RES.table)
}
\arguments{
\item{RES.table}{data frame of RES table from Sprint obtained from read.delim() function}
}
\description{
Assigns each RES from a Sprint output table a site name denoting position and type of change
}
