% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{run_ADARR}
\alias{run_ADARR}
\title{A function for annotating the Sprint results with genomic regions and repeats}
\usage{
run_ADARR(input.file.path, annotation.granges.path, complete.annotated.path)
}
\arguments{
\item{input.file.path}{The input file path}

\item{annotation.granges.path}{The path to output initial annotations}

\item{complete.annotated.path}{The path to output the complete annotations}
}
\description{
A function for annotating the Sprint results with genomic regions and repeats
}
