% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{run_ADARR_from_df}
\alias{run_ADARR_from_df}
\title{A function for annotating the Sprint results with genomic regions and repeats
This requires inputting a data frame rather than a file path}
\usage{
run_ADARR_from_df(
  input.sprint,
  annotation.granges.path,
  complete.annotated.path
)
}
\arguments{
\item{input.sprint}{The input file path}

\item{annotation.granges.path}{The path to output initial annotations}

\item{complete.annotated.path}{The path to output the complete annotations}
}
\description{
A function for annotating the Sprint results with genomic regions and repeats
This requires inputting a data frame rather than a file path
}
