% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{table_identified_genes}
\alias{table_identified_genes}
\title{A function for obtaining table of identified genes}
\usage{
table_identified_genes(input.sprint)
}
\arguments{
\item{input.sprint}{The Sprint data frame to input}
}
\description{
A function for obtaining table of identified genes
}
